fix: change NULL to nullptr, use lock_guard in send_to_client, use uint64_t for client_id

+++ include/client_info.h
@@ -8,7 +8,7 @@
  * @brief Holds all relevant information for a single connected client.
  */
 struct ClientInfo {
-	int client_id;
+	uint64_t client_id;
 	int socket_fd;
 	std::string ip_address;
 	int port;
